automation/discord-bot/log/requirements.txt
python-dotenv==1.0.0
boto3==1.28.40
discord.py==2.3.0
